change currentpage to int so we can go minus.

+++ PictureViewer01/Viewer.h
@@ -129,7 +129,7 @@ private:
 	float m_imageY{};
 	float m_scaleFactor = 1.0;
 
-	size_t m_currentPage = -1;
+	int m_currentPage = -1;
 	std::vector<std::shared_ptr<ZipFile>> m_zip_files;
 	std::vector<HGLOBAL> m_zip_globals;
